+++ activemq-unit-tests/src/test/java/org/apache/activemq/demo/SimpleProducer.java
@@ -28,13 +28,13 @@ package org.apache.activemq.demo;
 
 // START SNIPPET: demo
 
-import javax.jms.Connection;
-import javax.jms.ConnectionFactory;
-import javax.jms.Destination;
-import javax.jms.JMSException;
-import javax.jms.MessageProducer;
-import javax.jms.Session;
-import javax.jms.TextMessage;
+import jakarta.jms.Connection;
+import jakarta.jms.ConnectionFactory;
+import jakarta.jms.Destination;
+import jakarta.jms.JMSException;
+import jakarta.jms.MessageProducer;
+import jakarta.jms.Session;
+import jakarta.jms.TextMessage;
 import javax.naming.Context;
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
@@ -76,7 +76,7 @@ public final class SimpleProducer {
         destinationName = args[0];
         LOG.info("Destination name is " + destinationName);
         if (args.length == 2) {
-            numMsgs = (new Integer(args[1])).intValue();
+            numMsgs = new Integer(args[1]);
         } else {
             numMsgs = 1;
         }
